Rename commands to use underscores
Fixes #102
